extract get_changes into own module
