+++ app/models/journal.rb
@@ -30,6 +30,7 @@
 class Journal < ActiveRecord::Base
   self.table_name = 'journals'
 
+  include JournalChanges
   include JournalFormatter
   include FormatHooks
 
@@ -134,57 +135,6 @@ class Journal < ActiveRecord::Base
     if journable && !journable.changed?
       journable.touch
     end
-  end
-
-  def get_changes
-    return {} if data.nil?
-
-    if @changes.nil?
-      @changes = HashWithIndifferentAccess.new
-
-      if predecessor.nil?
-        @changes = data.journaled_attributes.select { |_, v| !v.nil? }
-                   .inject({}) { |h, (k, v)| h[k] = [nil, v]; h }
-      else
-        normalized_data = JournalManager.normalize_newlines(data.journaled_attributes)
-        normalized_predecessor_data = JournalManager.normalize_newlines(predecessor.data.journaled_attributes)
-
-        normalized_data.select do |k, v|
-          # we dont record changes for changes from nil to empty strings and vice versa
-          pred = normalized_predecessor_data[k]
-          v != pred && (v.present? || pred.present?)
-        end.each do |k, v|
-          @changes[k] = [normalized_predecessor_data[k], v]
-        end
-      end
-
-      @changes.merge!(get_association_changes predecessor, 'attachable', 'attachments', :attachment_id, :filename)
-      @changes.merge!(get_association_changes predecessor, 'customizable', 'custom_fields', :custom_field_id, :value)
-    end
-
-    @changes
-  end
-
-  def get_association_changes(predecessor, journal_association, association, key, value)
-    changes = {}
-    journal_assoc_name = "#{journal_association}_journals"
-
-    if predecessor.nil?
-      send(journal_assoc_name).each_with_object(changes) { |a, h| h["#{association}_#{a.send(key)}"] = [nil, a.send(value)] }
-    else
-      current = send(journal_assoc_name).map(&:attributes)
-      predecessor_attachable_journals = predecessor.send(journal_assoc_name).map(&:attributes)
-
-      merged_journals = JournalManager.merge_reference_journals_by_id current,
-                                                                      predecessor_attachable_journals,
-                                                                      key.to_s
-
-      changes.merge! JournalManager.added_references(merged_journals, association, value.to_s)
-      changes.merge! JournalManager.removed_references(merged_journals, association, value.to_s)
-      changes.merge! JournalManager.changed_references(merged_journals, association, value.to_s)
-    end
-
-    changes
   end
 
   def predecessor
